Update daily-hacks-for-smarter-living.md
Added the recording, a new image for the session card, and edited participant names. Also added an .ics line but I'm not sure if I'm doing that right... I saw it in the other session codes so felt like I needed to add it? Also, does this mean it's now automatically going from an 'upcoming session' to a 'past session' (based on the date)?